passing all user tests in standard api tests

+++ server/src/main/java/service/UserService.java
@@ -41,6 +41,10 @@ public class UserService {
     }
 
     public void logout(String authToken) throws DataAccessException {
+        AuthData auth = authDAO.getAuth(authToken);
+        if(auth == null) {
+            throw new DataAccessException("Auth token not found");
+        }
         authDAO.deleteAuth(authToken);
     }
